Add the New type of breadcrumb
