--- raaghu-elements/src/rds-advance-breadcrumb/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./rds-advance-breadcrumb";
