--- raaghu-elements/src/rds-advance-breadcrumb/rds-advance-breadcrumb.css
@@ -1,13 +1,0 @@
-.breadcrumb-background .breadcrumb-item {
-  padding: 0.5rem;
-  margin-right: 0.5rem;
-}
-
-.breadcrumb-background .breadcrumb-item.breadcrumb-round {
-  border-radius: 1.5rem;
-}
-.breadcrumb-separator {
-  display: flex;
-  align-items: center;
-  margin: 0 0.55rem;
-}
