+++ raaghu-elements/src/rds-breadcrumb/rds-breadcrumb.tsx
@@ -2,45 +2,55 @@ import React, { useEffect, useState } from "react";
 import "./rds-breadcrumb.css";
 import RdsIcon from "../rds-icon";
 
-export interface breadcrumbprop {
-    breadItems: any[];
+export interface BreadcrumbItem {
+    label: string;
+    id: number;
+    route?: string;
+    disabled?: boolean;
+    icon?: string;
+    iconFill?: boolean;
+    iconstroke?: boolean;
+    iconWidth?: string;
+    iconHeight?: string;
+    iconColor?: string;
+    active?: boolean;
 }
 
-const RdsBreadcrumb = (props: breadcrumbprop) => {
+export interface BreadcrumbProps {
+    breadItems: BreadcrumbItem[];
+    type: 'simple' | 'background';
+    shape?: 'round' | 'square';
+    separator?: '>' | '/' | '→' | '»' | '|' | '-';
+}
 
-    const [data, setdata] = useState(props.breadItems);
-    const length = data.length;
-
-    const onClickHandler = (key: any) => {
-        const tempData = data.map((item: any) => {
-            if (key === item.id && !item.disabled) {
-                return { ...item, active: !item.active };
-            } else {
-                return { ...item, active: false };
-            }
-        });
-
-        setdata(tempData);
-    };
+const RdsBreadcrumb =(props:BreadcrumbProps) => {
+    const [data, setData] = useState<BreadcrumbItem[]>(props.breadItems);
 
     useEffect(() => {
-        if (props.breadItems.length > 0) {
-            setdata(props.breadItems);
-        }
-
+        setData(props.breadItems);
     }, [props.breadItems]);
 
+    const onClickHandler = (key: number) => {
+        setData(data.map(item => ({
+            ...item,
+            active: key === item.id && !item.disabled ? !item.active : false,
+        })));
+    };
+
+    const shapeClass = props.shape ? `breadcrumb-${props.shape}` : '';
+    const noBgClass = props.shape ? `breadcrumb-${props.shape}-no-bg` : '';
+    const roundedClass = props.shape === 'round' ? 'rounded-5' : props.shape === 'square' ? 'rounded-2' : '';
+
     return (
-        <>
-            <nav aria-label="breadcrumb">
-                <ol className="breadcrumb m-0">
-                    {data?.map((breadItem, index) => {
-                        return index == 0 ? (
+        <nav aria-label="breadcrumb">
+            <ol className={`breadcrumb m-0 ${props.type === 'background' ? 'breadcrumb-background' : ''}`}>
+                {data.map((breadItem, index) => {
+                    const isLastItem = index === data.length - 1;
+
+                    return (
+                        <React.Fragment key={breadItem.id}>
                             <li
-                                key={breadItem.id}
-                                className={`breadcrumb-item  ${breadItem.active ? "active" : ""
-                                    } ${length - 1 === index ? "text-primary" : ""}`}
-                                id="breaditem1"
+                                className={`breadcrumb-item ${breadItem.active ? "active" : ""} ${isLastItem ? "text-primary" : ""} ${props.type === 'background' ? (isLastItem ? `${shapeClass} bg-primary-subtle` : 'bg-transparent') : ''} ${!isLastItem && breadItem.active ? noBgClass : ''} ${roundedClass}`}
                                 onClick={() => onClickHandler(breadItem.id)}
                             >
                                 {breadItem.icon && (
@@ -51,66 +61,27 @@ const RdsBreadcrumb = (props: breadcrumbprop) => {
                                             stroke={breadItem.iconstroke}
                                             width={breadItem.iconWidth}
                                             height={breadItem.iconHeight}
-                                            colorVariant={`${breadItem.active && breadItem.iconColor
-                                                }`}
+                                            colorVariant={breadItem.active ? breadItem.iconColor : ''}
                                         />
                                     </span>
                                 )}
                                 <a
-                                    // href={breadItem.route}
+                                    href={breadItem.route}
                                     className="text-decoration-none"
-                                    onClick={(event) => {
-                                        event.preventDefault();
-                                    }}
+                                    onClick={(e) => e.preventDefault()}
                                     aria-disabled="true"
                                 >
-                                    {(breadItem.label)}
+                                    {breadItem.label}
                                 </a>
                             </li>
-                        ) : (
-                            <li
-                                key={breadItem.id}
-                                className={`breadcrumb-items  ${breadItem.active ? "active" : ""
-                                    } ${length - 1 === index ? "text-primary" : ""}`}
-                                id={`breadcrumbItem+${breadItem.id}`}
-                                onClick={() => onClickHandler(breadItem.id)}
-                            >
-                                <span >
-                                    <RdsIcon
-                                        name="chevron_right"
-                                        stroke={true}
-                                        height="8px"
-                                        width="14px"
-                                    />
-                                </span>
-                                {breadItem.icon && (
-                                    <span className="ms-1 me-2">
-                                        <RdsIcon
-                                            name={breadItem.icon}
-                                            fill={breadItem.iconFill}
-                                            stroke={breadItem.iconstroke}
-                                            width={breadItem.iconWidth}
-                                            height={breadItem.iconHeight}
-                                            colorVariant={`${breadItem.active && breadItem.iconColor
-                                                }`}
-                                        />
-                                    </span>
-                                )}
-                                <a
-                                    // href={breadItem.route}
-                                    onClick={(event) => {
-                                        event.preventDefault();
-                                    }}
-                                    aria-disabled="true"
-                                >
-                                    {(breadItem.label)}
-                                </a>
-                            </li>
-                        );
-                    })}
-                </ol>
-            </nav>
-        </>
+                            {!isLastItem && (
+                                <li className="breadcrumb-separator">{props.separator}</li>
+                            )}
+                        </React.Fragment>
+                    );
+                })}
+            </ol>
+        </nav>
     );
 };
